# gcs/gcs: use sort.Slice instead of sort.Sort, remove uint64Slice